[optimize] build handleStatus function
