+++ src/components/login/SignInPopup.tsx
@@ -16,6 +16,7 @@ import { SignIn } from 'api/auth';
 import { setUserData } from 'redux/userSlice';
 // --- components ---
 import FormInput from 'components/form/FormInput';
+import { handleStatus } from 'utils/fetch';
 
 function SignInPopup() {
   const sliceDispatch = useDispatch();
@@ -64,7 +65,8 @@ function SignInPopup() {
       const variables = { email, password };
       try {
         const res = await SignIn(variables);
-        if (get(res, 'status') === 200) {
+
+        if (handleStatus(get(res, 'status', 0)) === 2) {
           const authToken = get(res, 'data.authToken');
           window.localStorage.setItem('authToken', authToken);
           setCookie('uid', res.data.userData.userId, { path: '/' });
@@ -83,8 +85,8 @@ function SignInPopup() {
               }
               handleClose();
             });
-        } else if (!isEmpty(get(res, 'response.data.message', ''))) {
-          setErrorMsg(get(res, 'response.data.message'));
+        } else if (handleStatus(get(res, 'status', 0)) === 4) {
+          setErrorMsg(get(res, 'data.message'));
         }
       } catch (error) {
         // console.log(error);
